Changed how weka was integrated by creating a source code version of the J48 classifier instead of loading the model file due to heap size constraints.
Hard coded 27 as the class index in one part of the PredictionModule because .classIndex() doesn't seem to give the right one.

Changed windowSize to tempWindow.size() in the calculate methods in PredictionModule to avoid having null pointer exceptions. But this seems to indicate that the number of sensor entries being processed is actually less than 100. Will fix this later.